refactor: move internal test files into test/source
